Refactor include paths to use geoobjects directory and enhance CMake configuration for Vcpkg integration
