+++ server/geoobjects/point.cpp
