+++ server/geoobjects/point.h
