+++ server/geoobjects/segment.cpp
