+++ server/geoobjects/segment.h
