+++ server/server.cpp
@@ -21,7 +21,7 @@
 #include <thread>
 
 // #include "db.h"
-#include "point.h"
+#include "geoobjects/point.h"
 #include "searchgraph.h"
 
 #include "server.h"
